requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-community
langchain-google-genai
langchain-groq
groq
langchainhub
duckduckgo-search
ddgs
faiss-cpu
sentence-transformers
pydantic
